+++ R/lasfilterdecimate.r
@@ -106,7 +106,7 @@ lasfilterdecimate.LAScatalog = function(las, algorithm)
   options <- list(need_buffer = FALSE, drop_null = TRUE, need_output_file = TRUE)
   output  <- catalog_apply(las, lasfilterdecimate, algorithm = algorithm, .options = options)
   output  <- unlist(output)
-  ctg     <- suppressMessages(suppressWarnings(catalog(output)))
+  ctg     <- suppressMessages(suppressWarnings(readLAScatalog(output)))
 
   opt_copy(ctg) <- las
   return(ctg)