+++ R/lasfiltersurfacepoints.r
@@ -76,7 +76,7 @@ lasfiltersurfacepoints.LAScatalog = function(las, res)
   options <- list(need_buffer = FALSE, drop_null = TRUE, need_output_file = TRUE)
   output  <- catalog_apply(las, lasfiltersurfacepoints, res = res, .options = options)
   output  <- unlist(output)
-  ctg     <- suppressMessages(suppressWarnings(catalog(output)))
+  ctg     <- suppressMessages(suppressWarnings(readLAScatalog(output)))
 
   opt_copy(ctg) <- las
   return(ctg)